+++ src/iter.rs
@@ -1,0 +1,234 @@
+//! Beap iterators.
+use super::Beap;
+use std::fmt;
+use std::iter::FusedIterator;
+
+impl<T> Beap<T> {
+    /// Returns an iterator visiting all values in the underlying vector, in
+    /// arbitrary order.
+    ///
+    /// # Examples
+    ///
+    /// Basic usage:
+    ///
+    /// ```
+    /// use beap::Beap;
+    /// let beap = Beap::from(vec![1, 2, 3, 4]);
+    ///
+    /// // Print 1, 2, 3, 4 in arbitrary order
+    /// for x in beap.iter() {
+    ///     println!("{}", x);
+    /// }
+    ///
+    /// assert_eq!(beap.into_sorted_vec(), vec![1, 2, 3, 4]);
+    /// ```
+    pub fn iter(&self) -> Iter<'_, T> {
+        Iter {
+            iter: self.data.iter(),
+        }
+    }
+
+    /// Clears the bi-parental heap, returning an iterator over the removed elements
+    /// in arbitrary order. If the iterator is dropped before being fully
+    /// consumed, it drops the remaining elements in arbitrary order.
+    ///
+    /// The returned iterator keeps a mutable borrow on the beap to optimize
+    /// its implementation.
+    ///
+    /// # Examples
+    ///
+    /// Basic usage:
+    ///
+    /// ```
+    /// use beap::Beap;
+    /// let mut beap = Beap::from([1, 3, 5]);
+    ///
+    /// assert!(!beap.is_empty());
+    ///
+    /// for x in beap.drain() {
+    ///     println!("{}", x);
+    /// }
+    ///
+    /// assert!(beap.is_empty());
+    /// ```
+    pub fn drain(&mut self) -> Drain<'_, T> {
+        self.height = 0;
+        Drain {
+            iter: self.data.drain(..),
+        }
+    }
+}
+
+impl<T> IntoIterator for Beap<T> {
+    type Item = T;
+    type IntoIter = IntoIter<T>;
+
+    /// Creates a consuming iterator, that is, one that moves each value out of
+    /// the beap in arbitrary order. The beap cannot be used
+    /// after calling this.
+    ///
+    /// # Examples
+    ///
+    /// Basic usage:
+    ///
+    /// ```
+    /// use beap::Beap;
+    /// let beap = Beap::from(vec![1, 2, 3, 4]);
+    ///
+    /// // Print 1, 2, 3, 4 in arbitrary order
+    /// for x in beap.into_iter() {
+    ///     // x has type i32, not &i32
+    ///     println!("{}", x);
+    /// }
+    /// ```
+    fn into_iter(self) -> IntoIter<T> {
+        IntoIter {
+            iter: self.data.into_iter(),
+        }
+    }
+}
+
+impl<'a, T> IntoIterator for &'a Beap<T> {
+    type Item = &'a T;
+    type IntoIter = Iter<'a, T>;
+
+    /// Returns an iterator visiting all values in the underlying vector, in
+    /// arbitrary order.
+    ///
+    /// # Examples
+    ///
+    /// Basic usage:
+    ///
+    /// ```
+    /// use beap::Beap;
+    /// let beap = Beap::from(vec![1, 2, 3, 4]);
+    ///
+    /// // Print 1, 2, 3, 4 in arbitrary order
+    /// for x in &beap {
+    ///     // x has type &i32
+    ///     println!("{}", x);
+    /// }
+    ///
+    /// assert_eq!(beap.into_sorted_vec(), vec![1, 2, 3, 4]);
+    /// ```
+    fn into_iter(self) -> Iter<'a, T> {
+        self.iter()
+    }
+}
+
+/// An iterator over the elements of a `Beap`.
+///
+/// This `struct` is created by [`Beap::iter()`]. See its
+/// documentation for more.
+///
+/// [`iter`]: Beap::iter
+#[derive(Clone)]
+pub struct Iter<'a, T: 'a> {
+    iter: std::slice::Iter<'a, T>,
+}
+
+impl<T: fmt::Debug> fmt::Debug for Iter<'_, T> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.debug_tuple("Iter").field(&self.iter.as_slice()).finish()
+    }
+}
+
+impl<'a, T> Iterator for Iter<'a, T> {
+    type Item = &'a T;
+
+    #[inline]
+    fn next(&mut self) -> Option<&'a T> {
+        self.iter.next()
+    }
+
+    #[inline]
+    fn size_hint(&self) -> (usize, Option<usize>) {
+        self.iter.size_hint()
+    }
+
+    #[inline]
+    fn last(self) -> Option<&'a T> {
+        self.iter.last()
+    }
+}
+
+impl<'a, T> DoubleEndedIterator for Iter<'a, T> {
+    #[inline]
+    fn next_back(&mut self) -> Option<&'a T> {
+        self.iter.next_back()
+    }
+}
+
+impl<T> FusedIterator for Iter<'_, T> {}
+
+/// An owning iterator over the elements of a `Beap`.
+///
+/// This `struct` is created by [`Beap::into_iter()`]
+/// (provided by the [`IntoIterator`] trait). See its documentation for more.
+///
+/// [`into_iter`]: Beap::into_iter
+/// [`IntoIterator`]: core::iter::IntoIterator
+#[derive(Clone)]
+pub struct IntoIter<T> {
+    iter: std::vec::IntoIter<T>,
+}
+
+impl<T: fmt::Debug> fmt::Debug for IntoIter<T> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.debug_tuple("IntoIter")
+            .field(&self.iter.as_slice())
+            .finish()
+    }
+}
+
+impl<T> Iterator for IntoIter<T> {
+    type Item = T;
+
+    #[inline]
+    fn next(&mut self) -> Option<T> {
+        self.iter.next()
+    }
+}
+
+impl<T> DoubleEndedIterator for IntoIter<T> {
+    #[inline]
+    fn next_back(&mut self) -> Option<T> {
+        self.iter.next_back()
+    }
+}
+
+impl<T> FusedIterator for IntoIter<T> {}
+
+/// A draining iterator over the elements of a `Beap`.
+///
+/// This `struct` is created by [`Beap::drain()`]. See its
+/// documentation for more.
+///
+/// [`drain`]: Beap::drain
+#[derive(Debug)]
+pub struct Drain<'a, T: 'a> {
+    iter: std::vec::Drain<'a, T>,
+}
+
+impl<T> Iterator for Drain<'_, T> {
+    type Item = T;
+
+    #[inline]
+    fn next(&mut self) -> Option<T> {
+        self.iter.next()
+    }
+
+    #[inline]
+    fn size_hint(&self) -> (usize, Option<usize>) {
+        self.iter.size_hint()
+    }
+}
+
+impl<T> DoubleEndedIterator for Drain<'_, T> {
+    #[inline]
+    fn next_back(&mut self) -> Option<T> {
+        self.iter.next_back()
+    }
+}
+
+impl<T> FusedIterator for Drain<'_, T> {}
